Reparacio del codi. No funcionava a causa de la SD. Que utilitza interrupcions que no es tenien en consideracio. Ara s'ha optat per una arquitectura una mica diferent. Les dades no es guarden de forma absolutament periodica, sino que s'esperen a que el APRS no estigui tx dades. La resta del sistema funciona de forma paralela com abans.

+++ include/BME_280.h
@@ -8,5 +8,5 @@ unsigned long BME_getPressure();
 char BME_getHumidity();
 float BME_getAltitude();
 void BME_print();
-void BME_log(unsigned long globalTime);
+void BME_log(uint32_t globalTime);
 #endif
